refactored App.js into different views
