+++ src/views/Debug.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Debug() {
+  return <div>Debug</div>;
+}
+
+export default Debug;
